refactor: 필터링 로직 커스텀훅으로 분리
