+++ src/pages/home/components/ProductList.tsx
@@ -1,7 +1,5 @@
-import { useEffect } from 'react';
-
-import { Product, Category } from '@/pages/home/types';
 import { useProductStore } from '@/store/useProductStore';
+import { useFilterList } from '@/pages/home/hooks/useFilterList';
 
 import ProductItem from './ProductItem';
 
@@ -9,25 +7,7 @@ export function ProductList() {
   const { productsData, currentCategory, filteredProducts, searchKeyword, setFilteredProducts } = useProductStore();
   const products = currentCategory || searchKeyword ? filteredProducts : productsData;
 
-  const getFilteredList = (products: Product[], filters: { category: Category | null; keyword: string }) => {
-    let filteredList = products;
-
-    if (filters.category) {
-      filteredList = filteredList.filter((product) => product.type === filters.category?.value);
-    }
-
-    if (filters.keyword) {
-      filteredList = filteredList.filter((product) => product.productName.includes(filters.keyword));
-    }
-
-    return filteredList;
-  };
-
-  useEffect(() => {
-    if (!currentCategory && !searchKeyword) return;
-    const filteredList = getFilteredList(productsData, { category: currentCategory, keyword: searchKeyword });
-    setFilteredProducts(filteredList);
-  }, [productsData, currentCategory, searchKeyword, setFilteredProducts]);
+  useFilterList(productsData, currentCategory, searchKeyword, setFilteredProducts);
 
   return (
     <>
